requirements.txt
llama-index
llama-index-llms-huggingface
llama-index-embeddings-huggingface
llama-index-embeddings-huggingface-api
python-dotenv
ipykernel
requests
matplotlib
utils
ollama
llama-index-llms-ollama
langchain
langchain-experimental
langchain-text-splitters
langchain-community
unstructured=0.7.12
chromadb